10th commit -> project prototyping

+++ src/main/java/com/soemoe/financetracker/utils/ValidationUtils.java
@@ -3,7 +3,7 @@ package com.soemoe.financetracker.utils;
 public class ValidationUtils {
     public static boolean isNumeric(String number) {
         try {
-            Integer.parseInt(number);
+            Double.parseDouble(number);
             return true;
         } catch (NumberFormatException e) {
             return false;
